added routing capabilities, working on user entity
