+++ client/src/pages/Login.tsx
@@ -1,0 +1,11 @@
+import { Stack } from "@chakra-ui/react";
+
+function Login() {
+    return (
+        <Stack>
+            <h1>This is my login page</h1>
+        </Stack>
+    )
+}
+
+export default Login;
